Revert "Flight Info make int field to double."
This reverts commit 7e108b9d598c810bfef8c9a7e182a8fecc063605.

+++ src/main/java/com/example/atc/data/FlightInfo.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class FlightInfo {
     private String flightName;
-    private double top;
-    private double left;
-    private double right;
-    private double down;
-    private double speed;
+    private int top;
+    private int left;
+    private int right;
+    private int down;
+    private int speed;
 }
